2wheels4u-master/2wheels4u-master/src/main/java/org/twowheels4u/security/UserDetailsServiceImpl.java
package org.twowheels4u.security;

import static org.springframework.security.core.userdetails.User.withUsername;

import lombok.RequiredArgsConstructor;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import org.twowheels4u.model.User;
import org.twowheels4u.service.UserService;

@Service
@RequiredArgsConstructor
public class UserDetailsServiceImpl implements UserDetailsService {

    private final UserService userService;

    @Override
    @Transactional
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user = userService.findByEmail(username);
        return withUsername(username)
                .password(user.getPassword())
                .roles(user.getRole().name())
                .build();
    }
}
